# tika-server-old/tika-server-spring/src/main/resources/api/tika-server-openapi.yaml
...
openapi: 3.0.2
...
info:
  contact:
    email: dev@tika.apache.org
  description: The Tika server implements [JAX-RS](http://en.wikipedia.org/wiki/JAX-RS) (Java API for RESTful Web Services) to provide web services according to the Representational State Transfer (REST) architectural style. This facilitates a wide varity oif operations and flexibility with regards to both client and server implementations. The officially supported Tika server implementation is packaged using the OpenAPI [jaxrs-cxf generator](https://openapi-generator.tech/docs/generators/jaxrs-cxf]. This work was tracked through [TIKA-3082](https://issues.apache.org/jira/browse/TIKA-3082).
    <b>N.B.</b> the OpenAPI version always tracks the underlying Tika version to remove uncertainty about which version of Tika is running within the server.
  license:
    name: Apache v2
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Tika JAX-RS Server
  version: 4.0.0-SNAPSHOT
servers:
  - description: Local development server
    url: 'https://localhost:9998'
...
paths:
...
  /async:
    post:
      description: POST a JSON request with fetch-emit tuples for asynchronous processing. The request must contain an emitter and a fetcherString key to fetch the inputStream. Optionally, it may contain metadata for pass-through.
      operationId: post_async
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                tuples:
                  type: array
                  items:
                    type: object
                    properties:
                      fetchKey:
                        type: object
                        properties:
                          fetcherName:
                            type: string
                          fetchKey:
                            type: string
                      emitKey:
                        type: object
                        properties:
                          emitterName:
                            type: string
                          emitKey:
                            type: string
                      metadata:
                        type: object
                        additionalProperties:
                          type: string
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  added:
                    type: integer
                    example: 1
          description: If successful, this operation returns HTTP status code 200 with status information.
        '429':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "throttled"
                  msg:
                    type: string
                  capacity:
                    type: integer
          description: Request throttled due to capacity limits.
        '400':
          description: Bad request - invalid fetcher or emitter.
        '500':
          description: An error occurred processing the call.
      summary: POST fetch-emit tuples for asynchronous processing
      tags:
      - Async Resource
...
  /language/stream:
    post:
      description: 'POST a UTF-8 text file to the LanguageIdentifier to identify its language. <b>NOTE</b>: This endpoint does not parse files.  It runs detection on a UTF-8 string.'
      operationId: post_language_stream
      responses:
        '200':
          content:
            text/plain:
              schema:
                example: 'en'
                type: string
          description: If successful, this operation returns HTTP status code 200, with the body being a string of the 2 character identified language.
        '500':
          description: An error occurred processing the call.
      summary: 'POST a UTF-8 text file to the LanguageIdentifier to identify its language.'
      tags:
      - Language Resource
    put:
      description: 'POST a UTF-8 text file to the LanguageIdentifier to identify its language. <b>NOTE</b>: This endpoint does not parse files.  It runs detection on a UTF-8 string.'
      operationId: put_language_stream
      responses:
        '200':
          content:
            text/plain:
              schema:
                example: 'en'
                type: string
          description: If successful, this operation returns HTTP status code 200, with the body being a string of the 2 character identified language.
        '500':
          description: An error occurred processing the call.
      summary: 'PUT a UTF-8 text file to the LanguageIdentifier to identify its language.'
      tags:
      - Language Resource
  /language/string:
    post:
      description: POST a text string to the LanguageIdentifier to identify its language.
      operationId: post_language_string
      responses:
        '200':
          content:
            text/plain:
              schema:
                example: 'en'
                type: string
          description: If successful, this operation returns HTTP status code 200, with the body being a string of the 2 character identified language.
        '500':
          description: An error occurred processing the call.
      summary: POST a text string to the LanguageIdentifier to identify its language.
      tags:
      - Language Resource
    put:
      description: PUT a text string to the LanguageIdentifier to identify its language.
      operationId: put_language_string
      responses:
        '200':
          content:
            text/plain:
              schema:
                example: 'en'
                type: string
          description: If successful, this operation returns HTTP status code 200, with the body being a string of the 2 character identified language.
        '500':
          description: An error occurred processing the call.
      summary: PUT a text string to the LanguageIdentifier to identify its language.
      tags:
      - Language Resource
...
  /status:
    get:
      description: GET the current server status including server ID, status, files processed, and restart count.
      operationId: get_status
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  server_id:
                    type: string
                  status:
                    type: string
                  millis_since_last_parse_started:
                    type: integer
                    format: int64
                  files_processed:
                    type: integer
                    format: int64
                  num_restarts:
                    type: integer
                    format: int32
          description: If successful, this operation returns HTTP status code 200 with server status information.
        '500':
          description: An error occurred processing the call.
      summary: GET the current server status
      tags:
      - Server Status
...